requirements.txt
pyyaml
xdg
